Fix priority not changing on reload
In 1.2.2 if a user changes the priority assigned to a vrrp-group and
reloads the configuration, the old priority will persist. This is
caused by saving the effective_priority in reset_vrrp_state. While
this makes sense for the case where a weighted interface or script is
used, it prevents to change the priority when sync groups are used (no
thread with vrrp_update_priority() is running in this case). Since
sync groups do not allow the use of weighted interfaces or scripts,
the effective priority is not copied over when they are used.

Original patch credit goes to John Southworth (Vyatta).